+++ examples/AspNet/Controllers/HomeController.cs
@@ -21,11 +21,13 @@ namespace Examples.AspNet.Controllers;
 public class HomeController : Controller
 {
     // For testing traditional routing. Ex: https://localhost:XXXX/
+    [HttpGet]
     public ActionResult Index()
     {
         return this.View();
     }
 
+    [HttpGet]
     [Route("about_attr_route/{customerId}")] // For testing attribute routing. Ex: https://localhost:XXXX/about_attr_route
     public ActionResult About(int? customerId)
     {