+++ src/OpenTelemetry.Instrumentation.AspNet.TelemetryHttpModule/TelemetryHttpModuleOptions.cs
@@ -51,16 +51,16 @@ public class TelemetryHttpModuleOptions
     /// <summary>
     /// Gets or sets a callback action to be fired when a request is started.
     /// </summary>
-    public Action<Activity, HttpContext> OnRequestStartedCallback { get; set; }
+    public Action<Activity, HttpContext>? OnRequestStartedCallback { get; set; }
 
     /// <summary>
     /// Gets or sets a callback action to be fired when a request is stopped.
     /// </summary>
-    public Action<Activity, HttpContext> OnRequestStoppedCallback { get; set; }
+    public Action<Activity, HttpContext>? OnRequestStoppedCallback { get; set; }
 
     /// <summary>
     /// Gets or sets a callback action to be fired when an unhandled
     /// exception is thrown processing a request.
     /// </summary>
-    public Action<Activity, HttpContext, Exception> OnExceptionCallback { get; set; }
+    public Action<Activity, HttpContext, Exception>? OnExceptionCallback { get; set; }
 }