+++ src/OpenTelemetry.Exporter.Stackdriver/Implementation/ResourceSemanticConventions.cs
@@ -1,0 +1,10 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+namespace OpenTelemetry.Exporter.Stackdriver.Implementation;
+
+internal class ResourceSemanticConventions
+{
+    public const string AttributeServiceName = "service.name";
+    public const string AttributeServiceVersion = "service.version";
+}