+++ src/OpenTelemetry.Exporter.Instana/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Update OpenTelemetry SDK version to `1.7.0`.
-  ([#1344](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1344))
+  ([#1486](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1486))
 * Drop support for .NET Framework 4.6.1.
   The lowest supported version is .NET Framework 4.6.2.
   ([#1050](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1050))