+++ src/OpenTelemetry.Exporter.Instana/CHANGELOG.md
@@ -5,8 +5,9 @@
 * Drop support for .NET Framework 4.6.1.
   The lowest supported version is .NET Framework 4.6.2.
   ([#1050](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1050))
-* Update OpenTelemetry SDK version to `1.8.1`.
-  ([#1668](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1668))
+
+* Updated OpenTelemetry core component version(s) to `1.9.0`.
+  ([#1888](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1888))
 
 ## 1.0.3
 
@@ -14,6 +15,7 @@ Released 2023-Feb-21
 
 * Fixes issue in span serialization process introduced in 1.0.2 version.
   ([#979](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/979))
+
 * Update OpenTelemetry SDK version to `1.3.2`.
   ([#917](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/917))
 
@@ -33,8 +35,10 @@ Released 2022-Nov-02
 
 * Instana span duration was not calculated correctly
   [376](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/376)
+
 * Application is crashing if environment variables are not defined
   [385](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/385)
+
 * Update OpenTelemetry SDK version to `1.3.1`.
   ([#749](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/749))